requirements.txt
# 운영 환경 패키지
fastapi==0.115.6
uvicorn[standard]==0.27.1
pydantic-settings~=2.7.1
pydantic~=2.10.5
sqlalchemy==2.0.27
email-validator==2.1.1
python-dotenv==1.0.1
bcrypt==4.1.2
requests==2.32.1
httpx==0.26.0
psycopg2-binary==2.9.10
alembic==1.11.1
pytz==2024.2
redis==5.2.1
aioredis==2.0.1
google-cloud-storage
openai
PyJWT==2.10.1
python-multipart
celery==5.3.1
importlib-metadata==4.13.0
prometheus-fastapi-instrumentator==7.0.2

# 테스트 및 개발 환경 패키지
pytest==7.4.3
pytest-subtests==0.7.0
pytest-asyncio==0.20.3
attrs==23.1.0
mypy==1.6.0
black==24.3.0
flower==2.0.1